[chore][pkg/stanza] Cleanup router operator files
